+++ src/MOI_wrapper/util.jl
@@ -112,3 +112,7 @@ function move_element_constraint(model)
         # @show typeof(constraint)
     end
 end
+
+function get_activator_internals(A, indices)
+    ActivatorConstraintInternals(A, indices[1] in indices[2:end], false, 0)
+end