+++ src/constraints/reified.jl
@@ -21,7 +21,6 @@ function init_constraint!(
             inner_constraint,
             inner_constraint.fct,
             inner_constraint.set;
-            active = false,
         )
         # map the bounds to the indicator constraint
         constraint.bound_rhs = inner_constraint.bound_rhs
@@ -37,7 +36,6 @@ function init_constraint!(
             anti_constraint,
             anti_constraint.fct,
             anti_constraint.set;
-            active = false,
         )
     end
     # still feasible
@@ -76,6 +74,7 @@ function prune_constraint!(
         !fix!(com, variables[rei_vidx], activate_off) && return false
         # 3
     elseif issetto(variables[rei_vidx], activate_on)
+        !activate_inner!(com, constraint) && return false
         return prune_constraint!(
             com,
             inner_constraint,
@@ -83,6 +82,7 @@ function prune_constraint!(
             inner_constraint.set,
         )
     elseif issetto(variables[rei_vidx], activate_off) && anti_constraint !== nothing
+        !activate_anti_inner!(com, constraint) && return false
         return prune_constraint!(
             com,
             anti_constraint,
@@ -198,117 +198,4 @@ function is_constraint_violated(
         )
     end
     return false
-end
-
-function update_best_bound_constraint!(
-    com::CS.CoM,
-    constraint::ReifiedConstraint,
-    fct::Union{MOI.VectorOfVariables,VAF{T}},
-    set::RS,
-    vidx::Int,
-    lb::Int,
-    ub::Int,
-) where {A,T<:Real,RS<:ReifiedSet{A}}
-    inner_constraint = constraint.inner_constraint
-    reified_vidx = constraint.indices[1]
-    search_space = com.search_space
-    reified_var = search_space[reified_vidx]
-    if inner_constraint.impl.update_best_bound
-        if CS.issetto(reified_var, Int(constraint.activate_on))
-            return update_best_bound_constraint!(
-                com,
-                inner_constraint,
-                inner_constraint.fct,
-                inner_constraint.set,
-                vidx,
-                lb,
-                ub,
-            )
-        else
-            # if not activated (for example in a different subtree) we reset the bounds
-            for rhs in constraint.bound_rhs
-                rhs.lb = typemin(Int64)
-                rhs.ub = typemax(Int64)
-            end
-        end
-    end
-    return true
-end
-
-function single_reverse_pruning_constraint!(
-    com::CoM,
-    constraint::ReifiedConstraint,
-    fct::Union{MOI.VectorOfVariables,VAF{T}},
-    set::RS,
-    var::Variable,
-    backtrack_idx::Int,
-) where {A,T<:Real,RS<:ReifiedSet{A}}
-    inner_constraint = constraint.inner_constraint
-    # the variable must be part of the inner constraint
-    if inner_constraint.impl.single_reverse_pruning &&
-       (var.idx != constraint.indices[1] || constraint.reified_in_inner)
-        single_reverse_pruning_constraint!(
-            com,
-            inner_constraint,
-            inner_constraint.fct,
-            inner_constraint.set,
-            var,
-            backtrack_idx,
-        )
-    end
-end
-
-function reverse_pruning_constraint!(
-    com::CoM,
-    constraint::ReifiedConstraint,
-    fct::Union{MOI.VectorOfVariables,VAF{T}},
-    set::RS,
-    backtrack_id::Int,
-) where {A,T<:Real,RS<:ReifiedSet{A}}
-    inner_constraint = constraint.inner_constraint
-    if inner_constraint.impl.reverse_pruning
-        reverse_pruning_constraint!(
-            com,
-            inner_constraint,
-            inner_constraint.fct,
-            inner_constraint.set,
-            backtrack_id,
-        )
-    end
-end
-
-function restore_pruning_constraint!(
-    com::CoM,
-    constraint::ReifiedConstraint,
-    fct::Union{MOI.VectorOfVariables,VAF{T}},
-    set::RS,
-    prune_steps::Union{Int,Vector{Int}},
-) where {A,T<:Real,RS<:ReifiedSet{A}}
-    inner_constraint = constraint.inner_constraint
-    if inner_constraint.impl.restore_pruning
-        restore_pruning_constraint!(
-            com,
-            inner_constraint,
-            inner_constraint.fct,
-            inner_constraint.set,
-            prune_steps,
-        )
-    end
-end
-
-function finished_pruning_constraint!(
-    com::CS.CoM,
-    constraint::ReifiedConstraint,
-    fct::Union{MOI.VectorOfVariables,VAF{T}},
-    set::RS,
-) where {A,T<:Real,RS<:ReifiedSet{A}}
-    inner_constraint = constraint.inner_constraint
-    if inner_constraint.impl.finished_pruning
-        finished_pruning_constraint!(
-            com,
-            inner_constraint,
-            inner_constraint.fct,
-            inner_constraint.set,
-        )
-    end
 end